AI/byzantine/pyproject.toml
[project]
name = "byzantine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.59.0",
    "flask>=3.1.1",
    "openai>=1.108.1",
    "python-dotenv>=1.1.1",
]
